+++ pkg/image/image.go
@@ -2,7 +2,7 @@
 //   Apptainer a Series of LF Projects LLC.
 //   For website terms of use, trademark policy, privacy policy and other
 //   project policies see https://lfprojects.org/policies
-// Copyright (c) 2018-2020, Sylabs Inc. All rights reserved.
+// Copyright (c) 2018-2022, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -370,6 +370,10 @@ func Init(path string, writable bool) (*Image, error) {
 	resolvedPath, err := ResolvePath(path)
 	if err != nil {
 		return nil, err
+	}
+
+	if !fs.IsReadable(resolvedPath) {
+		return nil, fmt.Errorf("%s is not readable by the current user, check permissions", resolvedPath)
 	}
 
 	img := &Image{